Models file containing models for lecture and speaker.
Current status is that they can be loaded using a store from a remote location.
The lecture object automatically initialises the speaker object it refers to.
It uses the config and initConfig combination to automatically generate
getters and setters for the objects.

Making it possible for a specific instance lecture1 to run:

alert(lecture1.getSpeaker().getDescription());

Code is documented